requirements.txt
amqp==5.1.1
asttokens==2.4.0
async-timeout==4.0.2
backcall==0.2.0
billiard==4.1.0
black==23.3.0
cachetools==5.3.1
celery==5.3.1
certifi==2023.5.7
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==2.2.1
comm==0.1.4
croniter==1.4.1
dask==2023.6.0
debugpy==1.8.0
decorator==5.1.1
dill==0.3.6
distributed==2023.6.0
docker==6.1.3
exceptiongroup==1.1.3
executing==1.2.0
fsspec==2023.6.0
google-api-core==2.11.1
google-auth==2.21.0
google-cloud-bigquery==3.11.3
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.56.0
grpcio-status==1.56.0
idna==3.4
importlib-metadata==6.7.0
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
Jinja2==3.1.2
jupyter_client==8.3.1
jupyter_core==5.3.2
kombu==5.3.1
locket==1.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-oneofschema==3.0.1
matplotlib-inline==0.1.6
msgpack==1.0.5
mypy-extensions==1.0.0
nest-asyncio==1.5.8
packaging==23.1
parso==0.8.3
partd==1.4.0
pathspec==0.11.1
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.8.0
prefect==0.15.9
prompt-toolkit==3.0.38
proto-plus==1.22.3
protobuf==4.23.3
psutil==5.9.5
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg-pool==3.1.7
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
python-box==7.0.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
pyzmq==25.1.1
redis==4.6.0
redis-pal==1.0.0
requests==2.26.0
rsa==4.9
six==1.16.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.19
stack-data==0.6.2
tblib==2.0.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
traitlets==5.10.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.6.1
zict==3.0.0
zipp==3.15.0
pandas
